feat: Trying a "nested" transaction
This is not exactly a typical nested transaction per se. Well, theore-
tically it is, since the utility methods are actually setting a save
point before consuming the connection, but the nested transaction is
using a separate connection, thus the "nested" transaction does not run
on the same connection. But I'm just trying out the waters...

Refs: #
Committed-by: Tasos Daris <tasos.daris@datawise.com>
Signed-off-by: Tasos Daris <tasos.daris@datawise.ai>